Add OrientedElement and DimensionalElement classes
