+++ src/compas_cadwork/elements/__init__.py
@@ -1,7 +1,9 @@
 from .beam import Beam
+from .dimensional_element import DimensionalElement
 from .element import Element
 from .element_type import ElementType
 from .factory import AnyElement
+from .oriented_element import OrientedElement
 from .panel import Panel
 from .wall import Wall
 
@@ -9,8 +11,10 @@ from .wall import Wall
 __all__ = [
     "AnyElement",
     "Beam",
+    "DimensionalElement",
     "Element",
     "ElementType",
+    "OrientedElement",
     "Panel",
     "Wall",
 ]
