+++ src/compas_cadwork/elements/factory.py
@@ -1,15 +1,19 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
+from typing import Final
 from typing import Literal
 from typing import TypeAlias
+from typing import get_args
 
 import attribute_controller as ac
 import element_controller as ec
 
 from .beam import Beam
+from .dimensional_element import DimensionalElement
 from .element import Element
 from .element_type import ElementType
+from .oriented_element import OrientedElement
 from .panel import Panel
 from .wall import Wall
 
@@ -18,37 +22,54 @@ if TYPE_CHECKING:
     from cadwork import ElementId
 
 
-_GenericElementTypes: TypeAlias = Literal[
+_BasicElementTypes: TypeAlias = Literal[
     ElementType.ADDITIONAL,
     ElementType.AUXILIARY,
     ElementType.CADWORK,
-    ElementType.CIRCULAR_AXIS,
-    ElementType.CONNECTOR_AXIS,
     ElementType.CONNECTOR_NODE,
     ElementType.CONTAINER,
-    ElementType.DIMENSION,
-    ElementType.DRILLING_AXIS,
-    ElementType.EAVE_AXIS,
     ElementType.EXPORT_SOLID,
     ElementType.EXPORT_SOLID_SCENE,
-    ElementType.FLOOR,
-    ElementType.GLOBAL_CUT,
-    ElementType.LINE,
     ElementType.NESTING_PARENT,
     ElementType.NONE,
     ElementType.NORMAL_NODE,
-    ElementType.OPENING,
-    ElementType.RECTANGULAR_AXIS,
-    ElementType.ROOF,
     ElementType.ROOM,
-    ElementType.ROTATION_ELEMENT,
     ElementType.SECTION_TRACE,
-    ElementType.SURFACE,
     ElementType.TEXT_DOCUMENT,
+]
+
+_OrientedElementTypes: TypeAlias = Literal[
+    ElementType.CIRCULAR_AXIS,
+    ElementType.CONNECTOR_AXIS,
+    ElementType.DIMENSION,
+    ElementType.DRILLING_AXIS,
+    ElementType.EAVE_AXIS,
+    ElementType.GLOBAL_CUT,
+    ElementType.LINE,
+    ElementType.RECTANGULAR_AXIS,
+    ElementType.ROTATION_ELEMENT,
+    ElementType.SURFACE,
     ElementType.WIRE_AXIS,
 ]
 
-AnyElement: TypeAlias = Element[_GenericElementTypes] | Beam | Panel | Wall
+_DimensionalElementTypes: TypeAlias = Literal[
+    ElementType.FLOOR,
+    ElementType.OPENING,
+    ElementType.ROOF,
+]
+
+AnyElement: TypeAlias = (
+    Element[_BasicElementTypes]
+    | OrientedElement[_OrientedElementTypes]
+    | DimensionalElement[_DimensionalElementTypes]
+    | Beam
+    | Panel
+    | Wall
+)
+
+_BASIC_ELEMENT_TYPES: Final = frozenset(get_args(_BasicElementTypes))
+_ORIENTED_ELEMENT_TYPES: Final = frozenset(get_args(_OrientedElementTypes))
+_DIMENSIONAL_ELEMENT_TYPES: Final = frozenset(get_args(_DimensionalElementTypes))
 
 
 def get_element_instance(cadwork_id: ElementId) -> AnyElement:
@@ -73,10 +94,22 @@ def get_element_instance(cadwork_id: ElementId) -> AnyElement:
         raise ValueError(f"Could not find a Cadwork element with ID #{cadwork_id}")
     raw_type = ac.get_element_type(cadwork_id)
     element_type = ElementType.from_cadwork(raw_type)
-    if element_type == ElementType.CIRCULAR_BEAM or element_type == ElementType.POLYGONAL_BEAM:
+
+    # Specific elements
+    if element_type in (ElementType.CIRCULAR_BEAM, ElementType.POLYGONAL_BEAM):
         return Beam(cadwork_id)
     if element_type == ElementType.PANEL:
         return Panel(cadwork_id)
     if element_type == ElementType.WALL:
         return Wall(cadwork_id)
-    return Element(cadwork_id)
+
+    # Generic elements
+    if element_type in _BASIC_ELEMENT_TYPES:
+        return Element(cadwork_id)
+    if element_type in _ORIENTED_ELEMENT_TYPES:
+        return OrientedElement(cadwork_id)
+    if element_type in _DIMENSIONAL_ELEMENT_TYPES:
+        return DimensionalElement(cadwork_id)
+
+    # Base case (this should never happen)
+    raise ValueError(f"Unmapped type {element_type.name!r} for Cadwork element with ID #{cadwork_id}")
