+++ tests/compas_cadwork/elements/test_panel.py
@@ -109,9 +109,3 @@ def test_creates_rectangular_panel(cadwork) -> None:
         cadwork.cadwork.point_3d(1.0, 0.0, 0.0),
         cadwork.cadwork.point_3d(0.0, -1.0, 0.0),
     )
-
-
-def test_repr(cadwork) -> None:
-    element = Panel(123)
-    cadwork.ac.get_name.return_value = "Something"
-    assert repr(element) == "Panel(id=123, name='Something')"
